Squash all artist-shuffle commits
Rename queue context for Top Songs from 'artist' to 'top-songs'.

Add ArtistWithSongs interface and associated mapping function

Add state functions to fetch artist with songs

Add hooks to use ArtistWithSongs

Add context menu to shuffle artist tracks.

TODO this is so very wrong. The useArtistWithSongs hook sometimes returns undefined (?) so I've opted to hide the shuffe options when this occurs.

It also populates ArtistWithSongs whenever the context menu is lauched, this should not happen. We should only fetch all the songs when the user actually clicks Shuffle. Clearly I have no idea how to structure a React app.

Replace useArtistWithSongs hook with useShuffleArtist which returns an async function to fetch the artist and update the queue.

Rename useShuffleArtist to usePlayArtist and have the return function accept a parameter to determine if the songs are shuffled or not.

Update ListPlayerControls to take onPlay and onShuffle callbacks instead of handling the queue directly

Add ListPlayerControls to ArtistView with options to shuffle and play artist songs

Rename OptionShuffleArtist to OptionPlayArtist and add a new properties to determine if the tracks are shuffled.

Update icons to match what is being used elsewhere in the app.

Use track ID instead of queue position to determine which song is currently playing in TitleTextSong views, should fix issue with play/shuffle artist function showing the wrong track playing

Use both queue position and track ID to identify playing track.

Construct a contextID unique to Top Songs so it won't match the Play/Shuffle contextId